SignalRService/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-blob==12.8.1
fastapi
llama_index
openai
python-dotenv
python-multipart
PyPDF2
